Undo extraction of inner classes. CXF has identical named types, and both ours and theirs are referenced in LiferayOAuth2DataProvider making it difficult to read
